Make parsing ticket status as utils
